# keydir: avoid defers