@prologic 2019-08-01
😱That's a significant performance improvement! Don't you think? i was aware there was a small cost with deferred(s) but not as much as 4-5% :D