@prologic 2019-08-01
If you're happy with this (I am!) I'm quite happy to merge this. Would you also consider doing the same for trie updates as well? Performance improvements at this stage of this library's development are very welcome! There are hopefully few critical bugs left!