@jsign 2019-08-01
TBH, quite surprised too, but when entering performance-mode it's better to test any assumption.

Sure, I'm ok with merging this, I'll take a look at trie updates too and issue another PR when ready.