@prologic 2019-08-01
In that case merged! I actually forked [trie](https://github.com/prologic/trie) so I'm also looking (_whilst also watching TV_) if I can apply the same optimisations as you've shown here :)