@jsign 2019-08-01
Cool!
I did my own experiment which I explained in a [new issue](https://github.com/derekparker/trie/issues/21).